[UX] Add Refresh button to all pages with lists #1997
