+++ frontend/package.json
@@ -17,17 +17,17 @@
     "generate-api": "npx @rtk-query/codegen-openapi openapi-config.ts"
   },
   "devDependencies": {
-    "@babel/cli": "^7.20.7",
-    "@babel/core": "^7.20.12",
+    "@babel/cli": "^7.25.9",
+    "@babel/core": "^7.26.0",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-transform-runtime": "^7.19.6",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@babel/register": "^7.18.9",
-    "@cfaester/enzyme-adapter-react-18": "^0.6.0",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
-    "@rtk-query/codegen-openapi": "^1.2.0",
+    "@babel/plugin-transform-runtime": "^7.25.9",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.25.9",
+    "@babel/preset-typescript": "^7.26.0",
+    "@babel/register": "^7.25.9",
+    "@cfaester/enzyme-adapter-react-18": "^0.8.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
+    "@rtk-query/codegen-openapi": "^2.0.0",
     "@svgr/webpack": "^6.5.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -37,7 +37,7 @@
     "@types/enzyme": "^3.10.18",
     "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.13",
-    "@types/node": "^22.8.4",
+    "@types/node": "^22.9.0",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/react-helmet": "^6.1.11",
@@ -46,8 +46,8 @@
     "@types/react-test-renderer": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^5.48.1",
     "@typescript-eslint/parser": "^5.48.1",
-    "@webpack-cli/serve": "^2.0.1",
-    "babel-loader": "^9.1.2",
+    "@webpack-cli/serve": "^2.0.5",
+    "babel-loader": "^9.2.1",
     "babel-preset-minify": "^0.5.2",
     "circular-dependency-plugin": "^5.2.2",
     "copy-webpack-plugin": "^11.0.0",
@@ -83,17 +83,17 @@
     "style-loader": "^4.0.0",
     "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.4",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1",
-    "webpack-merge": "^5.8.0",
+    "typescript": "^5.6.3",
+    "webpack": "^5.96.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0",
+    "webpack-merge": "^6.0.1",
     "webpack-nano": "^1.1.1"
   },
   "dependencies": {
     "@cloudscape-design/collection-hooks": "^1.0.55",
-    "@cloudscape-design/components": "^3.0.809",
-    "@cloudscape-design/design-tokens": "^3.0.47",
+    "@cloudscape-design/components": "^3.0.831",
+    "@cloudscape-design/design-tokens": "^3.0.49",
     "@cloudscape-design/global-styles": "^1.0.32",
     "@hookform/resolvers": "^2.9.10",
     "@reduxjs/toolkit": "^1.9.1",
@@ -117,7 +117,7 @@
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.27.0",
     "react-string-replace": "^1.1.1",
-    "redux": "^4.2.0",
+    "redux": "^5.0.1",
     "yup": "^0.32.11"
   }
 }
