+++ frontend/src/pages/Runs/List/hooks/index.ts
@@ -4,3 +4,4 @@ export * from './useDeleteRuns';
 export * from './useDisabledStatesForButtons';
 export * from './useEmptyMessages';
 export * from './useFilters';
+export * from './useRunsData';
