+++ frontend/src/services/auth.ts
@@ -1,6 +1,5 @@
 import { API } from 'api';
-import { createApi } from '@reduxjs/toolkit/query/react';
-import { fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
 import unauthorizedQueryHeaders from 'libs/fetchBaseQueryHeaders';
