+++ frontend/src/services/mainApi.ts
@@ -1,5 +1,4 @@
-import { fetchBaseQuery } from '@reduxjs/toolkit/dist/query/react';
-import { createApi } from '@reduxjs/toolkit/query/react';
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 
 import fetchBaseQueryHeaders from '../libs/fetchBaseQueryHeaders';
